files renamed and types specified
